Remove partial specification of FD coefficients
A slight optimisation in applying finite difference approximations to
odd derivatives because the coefficients are antisymmetric: one can
ignore the zero coefficient in the middle which saves a FLOP.
Previously we were open to this optimisation by allowing for partial
specification of finite difference coefficients in odd derivatives,
but we didn't actually use it and it was a headache thinking about
this optimisation prematurely. So this commit removes the interface
for the time being.